GA1: Move function to httpRequestUtil, update imports and remove requestUtil file
