+++ src/services/RegisterService.js
@@ -1,5 +1,4 @@
-import { createRequestData } from '../utils/requestUtil';
-import { httpRequest } from '../utils/httpUtil';
+import { httpRequest, createRequestData } from '../utils/httpRequestUtil';
 
 const REGISTER_URL = `${process.env.REACT_APP_REST_API}user/register`;
 
@@ -9,6 +8,7 @@ export const registerService = {
       return httpRequest(REGISTER_URL, () => createRequestData('POST', body));
     } catch (e) {
       console.error(e);
+      return e;
     }
   },
 };
