--- src/utils/httpUtil.js
@@ -1,5 +1,0 @@
-export const httpRequest = async (url, request) => {
-  const response = await fetch(url, request());
-  const responseData = await response.json();
-  return responseData;
-};
